refactor: Drop support for 'SectionStorage'
Instead of using session storage, the class was refactored to use
Ionic Storage. This means that the remaining files
(sectionStorage.service) act as a proxy for Ionic Storage
(storage.service). A full refactor should be scheduled to eliminate the
use of sectionStorage and replace it with the regular storage class.